--- a/pwnomcp/tools/pwndbg.py
+++ b/pwnomcp/tools/pwndbg.py
@@ -25,10 +25,17 @@ class PwndbgTools:
         """
         self.gdb = gdb_controller
         self.session = session_state
+        # Ensure GDB is initialized once lazily to avoid startup cost unless used
+        try:
+            self.gdb.initialize()
+        except Exception:
+            # Defer errors to first actual call to avoid failing construction
+            logger.exception("GDB initialize() failed during tool init; will retry on demand")
         
     def execute(self, command: str) -> Dict[str, Any]:
         """Execute arbitrary GDB/pwndbg command and return raw responses"""
         logger.info(f"Execute tool: {command}")
+        self.gdb.initialize()
         result = self.gdb.execute_command(command)
         self.session.update_state(result["state"])
         self.session.record_command(command, result)
@@ -37,6 +44,7 @@ class PwndbgTools:
     def set_file(self, binary_path: str) -> Dict[str, Any]:
         """Set the file to debug; return raw responses"""
         logger.info(f"Set file: {binary_path}")
+        self.gdb.initialize()
         result = self.gdb.set_file(binary_path)
         if result.get("success"):
             self.session.binary_path = binary_path
@@ -48,6 +56,7 @@ class PwndbgTools:
     def attach(self, pid: int) -> Tuple[Dict[str, Any], List[Dict[str, Any]]]:
         """Attach to an existing process; return raw responses"""
         logger.info(f"Attach to pid: {pid}")
+        self.gdb.initialize()
         result, context = self.gdb.attach(pid)
         if result.get("success"):
             self.session.pid = pid
@@ -58,6 +67,7 @@ class PwndbgTools:
     def run(self, args: str = "", start: bool = False) -> Dict[str, Any]:
         """Run the loaded binary; return raw responses"""
         logger.info(f"Run with args: '{args}'")
+        self.gdb.initialize()
         if not self.session.binary_loaded:
             return {"command": "-exec-run", "responses": [], "success": False, "state": self.gdb.get_state(), "error": "No binary loaded. Use set_file first."}
         result = self.gdb.run(args, start=start)
@@ -68,6 +78,7 @@ class PwndbgTools:
     def finish(self) -> Dict[str, Any]:
         """Run until current function finishes; return raw responses"""
         logger.info("Finish current function")
+        self.gdb.initialize()
         result = self.gdb.finish()
         self.session.update_state(result["state"])
         self.session.record_command(result.get("command", "-exec-finish"), result)
@@ -78,6 +89,7 @@ class PwndbgTools:
     def jump(self, locspec: str) -> Dict[str, Any]:
         """Jump to a specific location; return raw responses"""
         logger.info(f"Jump to {locspec}")
+        self.gdb.initialize()
         result = self.gdb.jump(locspec)
         self.session.update_state(result["state"])
         self.session.record_command(result.get("command", f"-exec-jump {locspec}"), result)
@@ -86,6 +98,7 @@ class PwndbgTools:
     def return_from_function(self) -> Dict[str, Any]:
         """Force return from current function; return raw responses"""
         logger.info("Force return from current function")
+        self.gdb.initialize()
         result = self.gdb.return_from_function()
         self.session.update_state(result["state"])
         self.session.record_command(result.get("command", "-exec-return"), result)
@@ -94,6 +107,7 @@ class PwndbgTools:
     def until(self, locspec: Optional[str] = None) -> Dict[str, Any]:
         """Run until a location or next source line; return raw responses"""
         logger.info(f"Until {locspec if locspec else '[next line]'}")
+        self.gdb.initialize()
         result = self.gdb.until(locspec)
         self.session.update_state(result["state"])
         self.session.record_command(result.get("command", "-exec-until"), result)
@@ -102,6 +116,7 @@ class PwndbgTools:
     def step_control(self, command: str) -> Dict[str, Any]:
         """Execute stepping commands (c, n, s, ni, si); return raw responses"""
         logger.info(f"Step control: {command}")
+        self.gdb.initialize()
         command_map = {
             "c": "continue",
             "n": "next",
@@ -132,14 +147,13 @@ class PwndbgTools:
     def get_context(self, context_type: str = "all") -> Dict[str, Any]:
         """Get debugging context (registers, stack, disassembly, etc.)"""
         logger.info(f"Get context: {context_type}")
+        self.gdb.initialize()
         if self.gdb.get_state() != "stopped":
             return {"command": f"context {context_type}", "responses": [], "success": False, "state": self.gdb.get_state(), "error": f"Cannot get context while inferior is {self.gdb.get_state()}"}
         if context_type == "all":
-            # Aggregate raw responses of each context call
-            aggregated = {"success": True, "state": self.gdb.get_state(), "contexts": {}}
-            for ctx_type in ["regs", "stack", "disasm", "code", "backtrace"]:
-                aggregated["contexts"][ctx_type] = self.gdb.get_context(ctx_type)
-            self.session.record_command("context all", aggregated)
+            # Prefer fast MI-based snapshot to avoid slow pwndbg rendering
+            aggregated = self.gdb.get_quick_context()
+            self.session.record_command("context all (quick)", aggregated)
             return aggregated
         else:
             result = self.gdb.get_context(context_type)
@@ -185,15 +199,26 @@ class PwndbgTools:
     def get_memory(self, address: str, size: int = 64, format: str = "hex") -> Dict[str, Any]:
         """Read memory at specified address; return raw responses"""
         logger.info(f"Read memory at {address}, {size} bytes as {format}")
+        self.gdb.initialize()
         if format == "hex":
-            cmd = f"hexdump {address} {size}"
+            # Use fast MI bytes read and return raw bytes; caller can format
+            result = self.gdb.read_memory_bytes(address, size)
+            self.session.record_command(f"-data-read-memory-bytes {address} {size}", result)
+            return result
         elif format == "string":
+            # For C-string, classic command is fine (needs pwndbg/pretty print)
             cmd = f"x/s {address}"
+            result = self.gdb.execute_command(cmd)
+            self.session.record_command(cmd, result)
+            return result
         else:
-            cmd = f"x/{size}b {address}"
-        result = self.gdb.execute_command(cmd)
-        self.session.record_command(cmd, result)
-        return result
+            # Generic bytes read via MI grid as fallback
+            word_size = 1
+            nr_rows = size
+            nr_cols = 1
+            result = self.gdb.read_memory_mi(address, "x", word_size, nr_rows, nr_cols)
+            self.session.record_command(f"-data-read-memory {address} x {word_size} {nr_rows} {nr_cols}", result)
+            return result
         
     def get_session_info(self) -> Dict[str, Any]:
         """Get current session information (no GDB sync/parsing)"""