Performance great upgrade.
* init-time 10 -> 1 seconds
* execution-time 10 -> 1 seconds

got ride of useless redundant things written by Claude that slowed things down.

Todo:
- [ ] context optimization